src/requirements.txt
alembic==1.14.1
APScheduler==3.11.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==0.21.0
pytz==2024.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
Werkzeug==3.1.3
